Modularize `app()` to enable testing routes
