+++ server.js
@@ -1,12 +1,5 @@
-const express = require('express');
-const app = express();
+const app = require('./app');
 const path = require('path');
-const bodyParserJSON = require('body-parser').json;
-
-const gamesRoutes = require('./api/routes/games');
-const priceAlertsRoutes = require('./api/routes/priceAlerts');
-const blacklistRoutes = require('./api/routes/blacklist');
-
 
 app.set('port', (process.env.PORT || 3001));
 
@@ -20,8 +13,3 @@ if (process.env.NODE_ENV === 'production') {
 app.listen(app.get('port'), function () {
     console.log(`Find the server at: http://localhost:${app.get('port')}/`); // eslint-disable-line no-console
 });
-
-app.use(bodyParserJSON());
-gamesRoutes(app);
-priceAlertsRoutes(app);
-blacklistRoutes(app);
